+++ web/src/settings_panel_menu.ts
@@ -9,13 +9,9 @@ import {$t, $t_html} from "./i18n.ts";
 import * as keydown_util from "./keydown_util.ts";
 import * as popovers from "./popovers.ts";
 import * as scroll_util from "./scroll_util.ts";
+import {redraw_all_bots_list, redraw_your_bots_list} from "./settings_bots.ts";
 import * as settings_sections from "./settings_sections.ts";
-import {
-    redraw_active_users_list,
-    redraw_all_bots_list,
-    redraw_deactivated_users_list,
-    redraw_your_bots_list,
-} from "./settings_users.ts";
+import {redraw_active_users_list, redraw_deactivated_users_list} from "./settings_users.ts";
 import * as util from "./util.ts";
 
 export let normal_settings: SettingsPanelMenu;
