+++ web/src/settings_sections.ts
@@ -4,6 +4,7 @@ import * as alert_words_ui from "./alert_words_ui.ts";
 import * as attachments_ui from "./attachments_ui.ts";
 import * as blueslip from "./blueslip.ts";
 import * as settings_account from "./settings_account.ts";
+import * as settings_bots from "./settings_bots.ts";
 import * as settings_emoji from "./settings_emoji.ts";
 import * as settings_exports from "./settings_exports.ts";
 import * as settings_folders from "./settings_folders.ts";
@@ -65,7 +66,7 @@ export function initialize(): void {
 
     // org
     load_func_dict.set("org_misc", settings_org.set_up);
-    load_func_dict.set("org_bots", settings_users.set_up_bots);
+    load_func_dict.set("org_bots", settings_bots.set_up_bots);
     load_func_dict.set("org_users", settings_users.set_up_humans);
     load_func_dict.set("emoji-settings", settings_emoji.set_up);
     load_func_dict.set("default-channels-list", settings_streams.set_up);
