+++ web/tests/example4.test.cjs
@@ -59,6 +59,7 @@ const {run_test, noop} = require("./lib/test.cjs");
 const activity_ui = mock_esm("../src/activity_ui");
 const message_live_update = mock_esm("../src/message_live_update");
 const pm_list = mock_esm("../src/pm_list");
+const settings_bots = mock_esm("../src/settings_bots");
 const settings_users = mock_esm("../src/settings_users");
 const user_profile = mock_esm("../src/user_profile");
 
@@ -88,7 +89,7 @@ run_test("add users with event", ({override}) => {
 
     // We need to override a stub here before dispatching the event.
     // Keep reading to see how overriding works!
-    override(settings_users, "redraw_all_bots_list", noop);
+    override(settings_bots, "redraw_all_bots_list", noop);
     override(activity_ui, "check_should_redraw_new_user", noop);
     // Let's simulate dispatching our event!
     server_events_dispatch.dispatch_normal_event(event);
@@ -140,7 +141,7 @@ run_test("update user with event", ({override}) => {
     override(message_live_update, "update_user_full_name", noop);
     override(pm_list, "update_private_messages", noop);
     override(settings_users, "update_user_data", noop);
-    override(settings_users, "update_bot_data", noop);
+    override(settings_bots, "update_bot_data", noop);
     override(user_profile, "update_profile_modal_ui", noop);
 
     // Dispatch the realm_user/update event, which will update
